requirements.txt

streamlit==1.12.0
base64=3.3

gunicorn==21.2.0
Pillow>=8.0.0
protobuf==3.20.3


tensorflow>=2.13.0,<=2.15.0
keras==2.13.1
keras_applications==1.0.8

pyautogui==0.9.54
absl-py==1.4.0
altair==4.2.2

argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3

attrs==23.1.0
Babel==2.12.1

bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0

cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
flatbuffers==23.5.26

fqdn==1.5.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.33
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0

helper==2.5.0
idna==3.4
ipytablewidgets==0.3.1

ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4


kiwisolver==1.4.5
libclang==16.0.6
lz4==4.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1

nest-asyncio==1.5.7
nltk==3.2.4

notebook_shim==0.2.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5

platformdirs==3.10.0
preprocessing==0.1.13
prometheus-client==0.17.1
prompt-toolkit==3.0.39

psutil==5.9.5
pure-eval==0.2.2

pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2023.3

PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2

rsa==4.9
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sphinx-rtd-theme==0.2.4
stack-data==0.6.2

tensorboard-data-server==0.7.1

tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tokenizer==3.4.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0

traitlets==5.9.0
traittypes==0.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
widgetsnbextension==4.0.8
wrapt==1.15.0
